Skip calc if resolution is zero.
This isn't something we should really be following for the difference
series. We don't need to pad this yet, we only need to pad when we are
trying to add data series.

We should also have a job that checks each sensor and the measurements
to see if the resolution of the measurements matches the resolution as
stored on the sensor and update it.