UI part.
Add a drop down to allow to specify the date format in the file.

A bit bodgy, need to refactor existing fileupload to allow additional
fields to be passed. Also 'selected' format is stored in a hidden field
to allow the existing upload code to be used. Might be better to
refactor that to suck info from the app/local state to populate the form
POST. something for another day.