Scripts etc to backup the dse cluster to S3.
cronjob runs on kixi-app0 which runs a script on each dse node in
parallel.

Those dse scripts take a snapshot and then sync *all* snapshots to S3.

This means that we get enventually consistent snapshots archived in S3,
we need to decide how often we want to do this. We can run the script
manually whenever we want and it should work.